Fix Bags

+++ app/src/main/java/com/example/myweather/Model2/Coord.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord(
     @SerializedName("lat")
-    val lat: Int = 0,
+    val lat: Double = 0.0,
     @SerializedName("lon")
     val lon: Double = 0.0
 )
